// guibiometricos/src/app/module/testlifecredencial/model/testlifecredencial.ts
export class   Testlifecredencial  {




  fotoPruebaVida?: string;
  fotoCredencial?: string;
  curp?: string;
  username ?: string;
  nombreUsuario?: string;
  rfc ?: string;
  cveIdUsuario?: string;




  claveElector?: boolean;
	anioRegistro?: boolean;
	apellidoPaterno?: boolean;
	anioEmision?: boolean;
	numeroEmisionCredencial?: boolean;
	nombre?: boolean;
	curpvalida?: boolean;
	apellidoMaterno?: boolean;
	ocr?: boolean;
  resultado?: string;

  cofirmavida?: boolean;

   desCurpCredencial?: string;
   desNombre?: string;
   desPrimerApellido?: string;
   desSegundoApellido?: string;
   desClaveElector?: string;

   desNumeroEmision?: string;
   desAnioRegistro?: string;
   desAnioEmision?: string;
   latitud ?: string;
   longitud ?: string;

   estado ?: number;

  constructor(){
  }

}